Moved company list to its own file.

+++ populate.js
@@ -5,7 +5,9 @@ const productsModel = require('./models/product');
 // For prepopulating the api with data.
 const start = async () => {
   try {
-    await productsModel.deleteMany({}); // Deletes all the currert data in the database
+    if (process.argv[2] === 'clean') {
+      await productsModel.deleteMany({}); // Deletes all the currert data in the database
+    }
     await productsModel.create(data); // Populates the database.
     console.log('Sucessfully poupulated the database!');
     process.exit(0);
@@ -16,4 +18,3 @@ const start = async () => {
 };
 
 start();
-// Setup args for deleting database before populating
